+++ client/src/style/HeaderFeedStyled.style.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 const HeaderFeedStyled = styled.header`
   display: flex;
   flex-direction: row-reverse;
-  background-color: blue;
+  background-color: ${props => props.theme.color};
   img {
     width: 40px;
     border-radius: 50%;
