+++ client/src/style/LogginSpaceStyled.style.js
