+++ client/src/variables.js
@@ -1,0 +1,4 @@
+export default {
+  color: "blue"
+}
+
